+++ client/src/pages/dashboard.jsx
@@ -1,0 +1,11 @@
+function Dashboard() {
+    return (
+        <div className="dashboard">
+            <h1>Hi, I'm your <span>Dashboard 🤗</span></h1>
+
+
+        </div>
+    )
+}
+
+export default Dashboard;