+++ client/src/pages/login.jsx
@@ -1,11 +1,14 @@
 import FormContainer from "../components/FormContainer";
 import InputField from "../components/InputField";
 import Button from "../components/Button";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import { useState } from "react";
+import toast from "react-hot-toast";
+import { login } from "../services/userServces";
 
 
 function Login() {
+    const navigate = useNavigate();
     // Handle form data
     const [formData, setFormData] = useState({
         email: "",
@@ -23,6 +26,8 @@ function Login() {
         })
 
     }
+    // Loading
+    const [loading, setLoading] = useState(false);
 
     // Handle error feedbacks
     const validate = () => {
@@ -50,18 +55,51 @@ function Login() {
 
     }
     // Handle form submit
-    const handleSubmit = (e) => {
+    const handleSubmit = async (e) => {
         e.preventDefault();
         const isValid = validate();
         if (!isValid) return;
-        console.log("form Submitted");
-    }
+
+        setLoading(true);
+        // Submit user data to the backend api
+        try {
+            const response = await login(formData);
+            toast.success(response.data.message);
+
+            setFormData({
+                email: "",
+                password: ""
+            });
+            setFieldErrors({});
+            // Redirect to dashboard
+            setTimeout(() => {
+                navigate('/dashboard');
+            }, 2000);
+        } catch (error) {
+            const backendMessage = error?.response?.data;
+            if (backendMessage?.errors) {
+                // Handle validation errors from backend
+                const mappedErrors = {};
+                backendMessage.errors.forEach((err) => {
+                    mappedErrors[err.field] = err.message;
+                    
+                }); 
+                setFieldErrors(mappedErrors);
+            } else {
+                toast.error(backendMessage?.message || "Internal server error. Please try again later.");
+            }
+
+        } finally {
+            setLoading(false);
+        }
+
+    };
 
 
     return (
         <FormContainer title="Welcome back!" subtitle="Enter your details to continue">
             <form onSubmit={handleSubmit} noValidate>
-               
+
                 <InputField
                     label="Email Address"
                     type="email"
@@ -81,7 +119,11 @@ function Login() {
                     error={fieldErrors.password}
 
                 />
-                <Button text="Login" type="submit" />
+                <Button
+                    text={loading ? "..." : "Login"}
+                    type="submit"
+                    disabled={loading}
+                />
                 <div className="formFooter">
                     <p> Don't have an account? </p>
                     <Link to="/create-account">Create Account</Link>