+++ client/src/services/userServces.js
@@ -1,0 +1,12 @@
+import api from "../api/axios";
+
+
+// Accessing backend create account api
+export const createAccount = (formData) => {
+    return api.post("/users/auth/create-account", formData)
+};
+
+// Accessing backend login api
+export const login = (formData) => {
+    return api.post("/users/auth/login", formData)
+};