+++ CHANGELOG.md
@@ -1,16 +1,23 @@
 # Changelog
 
-## v0.1.0 - Account Creation and Login
+## [ 0.1.0 ] - 28 July 2026 
 ### Added
+### Account Creation and Login
 - Create-account endpoint with password hashing 
 - Login endpoint with hashed password comparison
 - Input validation and sanitization via middleware
 - Database-level schema validation as a second layer of defense
 
-## v0.2.0 - Frontend for Create Acoount and Login
+## [ 0.2.0 ] - 29 July 2026 
 ### Added
+### Frontend for Create Acoount and Login
 - Create-account and login pages using React js with vanilla css
 - Fixing words missing in backend
 
+
+## [ 0.3.1 ] - 29 July 2026
+### Added
+### Integrated frontend with backend API
+- Connect frontend (create account & login forms) with backend api using axios for full functioning